+++ Cargo.toml
@@ -133,6 +133,7 @@ gpu_witness_eval_generator = { path = "./gpu_witness_eval_generator" }
 gpu_prover = { path = "./gpu_prover" }
 
 serde = { version = "1", default-features = false, features = ["derive", "alloc"] }
+clap = { version = "4.5.21", features = ["derive"] }
 # rand = {version = "0.8", default-features = false, features = ["std_rng"] }
 rand = { version = "0.9", default-features = false }
 unroll = "0.1"
