+++ tools/cli/src/lib.rs
@@ -2,19 +2,6 @@
 #![allow(incomplete_features)]
 #![feature(generic_const_exprs)]
 
-pub mod generate_constants;
 pub mod prover_utils;
 pub mod setup;
 pub mod vk;
-
-use clap::ValueEnum;
-use serde::{Deserialize, Serialize};
-
-#[derive(Clone, Debug, ValueEnum, Serialize, Deserialize, PartialEq, Eq)]
-pub enum Machine {
-    Standard,
-    Reduced,
-    ReducedLog23,
-    // Final reduced machine, used to generate a single proof at the end.
-    ReducedFinal,
-}
