+++ tools/generator/Cargo.toml
@@ -10,7 +10,7 @@ keywords.workspace = true
 categories.workspace = true
 
 [dependencies]
-clap = { version = "4.5.21", features = ["derive"] }
+clap = { workspace = true }
 prover = { workspace = true, features = ["prover"] }
 serde = { workspace = true }
 serde_json = { version = "*" }
